+++ fresh.gen.ts
@@ -15,6 +15,8 @@ import * as $DarkModeToggle from "./islands/DarkModeToggle.tsx";
 import * as $GitHubTab from "./islands/GitHubTab.tsx";
 import * as $Header from "./islands/Header.tsx";
 import * as $HeaderControls from "./islands/HeaderControls.tsx";
+import * as $LabelDropdown from "./islands/LabelDropdown.tsx";
+import * as $LabelManager from "./islands/LabelManager.tsx";
 import * as $Statistics from "./islands/Statistics.tsx";
 import * as $modals_CardModal from "./islands/modals/CardModal.tsx";
 import * as $modals_ChangelogModal from "./islands/modals/ChangelogModal.tsx";
@@ -41,6 +43,8 @@ const manifest = {
     "./islands/GitHubTab.tsx": $GitHubTab,
     "./islands/Header.tsx": $Header,
     "./islands/HeaderControls.tsx": $HeaderControls,
+    "./islands/LabelDropdown.tsx": $LabelDropdown,
+    "./islands/LabelManager.tsx": $LabelManager,
     "./islands/Statistics.tsx": $Statistics,
     "./islands/modals/CardModal.tsx": $modals_CardModal,
     "./islands/modals/ChangelogModal.tsx": $modals_ChangelogModal,